# AXMARIL/static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "sean",
        "version": "1.0.0",
        "title": "Seans-Python3-Flask-Rest-Boilerplate",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "/api/v1"
        },
        {
            "url": "/api/v2"
        }
    ],
    "security": [
        {
            "BearerAuth": []
        }
    ],
    "paths": {
...
        "/secret/safe/create": {
            "post": {
                "tags": [
                    "Safe"
                ],
                "summary": "Create a new coffre (safe)",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CoffreCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Coffre created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CoffreResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/secret/safe/delete": {
            "delete": {
                "tags": [
                    "Safe"
                ],
                "summary": "Delete a coffre",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CoffreDelete"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Coffre deleted successfully"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/secret/user/safe/all": {
            "get": {
                "tags": [
                    "Safe"
                ],
                "summary": "Get list of coffres for user",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of coffres",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CoffreListResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter 'Bearer <token>'"
            }
        },
        "schemas": {
...
            "CoffreCreate": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "required": [
                    "name"
                ]
            },
            "CoffreResponse": {
                "type": "object",
                "properties": {
                    "safe_id": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
            "CoffreDelete": {
                "type": "object",
                "properties": {
                    "safe_name": {
                        "type": "string"
                    }
                },
                "required": [
                    "safe_name"
                ]
            },
            "CoffreListResponse": {
                "type": "object",
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CoffreItem"
                        }
                    },
                    "total_count": {
                        "type": "integer"
                    }
                }
            },
            "CoffreItem": {
                "type": "object",
                "properties": {
                    "safe_id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "owner_uid": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}